Add special imports utility module
Collates the special import logic for several packages in one place.
